Easier loader configurations
