+++ src/entrypoints/array.ts
@@ -1,0 +1,3 @@
+import { init_from_snippet } from '../posthog-core'
+
+init_from_snippet()
