+++ src/entrypoints/recorder-v2.ts
@@ -1,0 +1,1 @@
+export * from './recorder'
