+++ src/entrypoints/surveys-preview.mts
@@ -1,0 +1,1 @@
+export { renderSurveysPreview, renderFeedbackWidgetPreview } from '../extensions/surveys'
