--- src/loader-exception-autocapture.ts
@@ -1,9 +1,0 @@
-import { extendPostHog } from './extensions/exception-autocapture'
-
-import { window } from './utils/globals'
-
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-window.extendPostHogWithExceptionAutoCapture = extendPostHog
-
-export default extendPostHog
