--- src/loader-globals.ts
@@ -1,3 +1,0 @@
-import { init_from_snippet } from './posthog-core'
-
-init_from_snippet()
